+++ bank-feeds/src/codatbankfeeds/models/operations/create_bank_account_mapping.py
@@ -5,42 +5,8 @@ import dataclasses
 import requests as requests_http
 from ..shared import bankfeedaccountmappingresponse as shared_bankfeedaccountmappingresponse
 from ..shared import errormessage as shared_errormessage
-from codatbankfeeds import utils
-from dataclasses_json import Undefined, dataclass_json
+from ..shared import zero as shared_zero
 from typing import Optional
-
-
-@dataclass_json(undefined=Undefined.EXCLUDE)
-@dataclasses.dataclass
-class CreateBankAccountMappingBankFeedAccountMapping:
-    r"""A bank feed connection between a source account and a target account."""
-    feed_start_date: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('feedStartDate'), 'exclude': lambda f: f is None }})
-    r"""In Codat's data model, dates and times are represented using the <a class=\\"external\\" href=\\"https://en.wikipedia.org/wiki/ISO_8601\\" target=\\"_blank\\">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
-
-    ```
-    2020-10-08T22:40:50Z
-    2021-01-01T00:00:00
-    ```
-
-
-
-    When syncing data that contains `DateTime` fields from Codat, make sure you support the following cases when reading time information:
-
-    - Coordinated Universal Time (UTC): `2021-11-15T06:00:00Z`
-    - Unqualified local time: `2021-11-15T01:00:00`
-    - UTC time offsets: `2021-11-15T01:00:00-05:00`
-
-    > Time zones
-    > 
-    > Not all dates from Codat will contain information about time zones.  
-    > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
-    """
-    source_account_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('sourceAccountId'), 'exclude': lambda f: f is None }})
-    r"""Unique ID for the source account"""
-    target_account_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('targetAccountId') }})
-    r"""Unique ID for the target account"""
-    
-
 
 
 @dataclasses.dataclass
@@ -49,7 +15,7 @@ class CreateBankAccountMappingRequest:
     r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
     r"""Unique identifier for a connection."""
-    request_body: Optional[CreateBankAccountMappingBankFeedAccountMapping] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
+    zero: Optional[shared_zero.Zero] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})