+++ bank-feeds/RELEASES.md
@@ -739,3 +739,13 @@ Based on:
 - [python v4.1.0] bank-feeds
 ### Releases
 - [PyPI v4.1.0] https://pypi.org/project/codat-bankfeeds/4.1.0 - bank-feeds
+
+## 2023-10-30 01:26:43
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.1] bank-feeds
+### Releases
+- [PyPI v4.1.1] https://pypi.org/project/codat-bankfeeds/4.1.1 - bank-feeds